+++ src/index.ts
@@ -1,0 +1,9 @@
+import { GCodePreview } from './gcode-preview';
+import { GCodeParser } from './gcode-parser';
+
+export { 
+    GCodePreview,
+    // GCodePreview as init,
+    GCodePreview as create,
+    GCodeParser
+}